change how strategy is approached
now, it's metadata for the experiment.. so you have "raw" or "Reg" or
whatever on the level of the 'experiment' object. this gets rid of there
being both 'raw_profiles' and 'reg_profiles' in the experiment object,
and only trimmed/untrimmed profiles. can still compare raw and reg by
concatenating the DataArrays into a DataSet